Added a quick regex check for quicker match verification
Made significant changes to the the check_rename function. It now performs a regex against the current filename and the proposed filename to see if a case-insensitive match of S##E## is present and notifies the user if the number of matched files is equal to the total number of files. This helps identify mismatched episodes within a single season.

Downside: the check_rename and rename_file functions are now drastically different, yet should be emulating the same renaming process. Should be refactored later to make them similar, or possible part of the same function.